+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/impl/projection/impl/column/ExpectedColumnProjection.java
@@ -15,19 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.impl;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.impl.column;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedSegment;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedDelimiterSegment;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.owner.ExpectedTableOwner;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.ExpectedProjection;
 
+import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 
 @Getter
 @Setter
-public final class ExpectedShorthandProjection extends AbstractExpectedSegment implements ExpectedProjection {
+public final class ExpectedColumnProjection extends AbstractExpectedDelimiterSegment implements ExpectedProjection {
+    
+    @XmlAttribute
+    private String name;
+    
+    @XmlAttribute
+    private String alias;
     
     @XmlElement(name = "owner")
     private ExpectedTableOwner owner;
